Update content_types_script.html
Replaced the JS with an array variable, then used a loop to cycle through them based upon click detection. I also added a reset JS button function instead of needing to reload the page. This JS is more efficient, and does not require the reloading of the page just to reset the  the span elements, and re-display the buttons. The reset function recycles the const variable calling the variable within the function using the console.log feature.